Downloads movie data from twitter
Got through the whole list of movies, fixed  some spelling mistakes in the
movie list itself. Lots of error and if one does occur it just moves onto
the next  movie. given 5 search results from wikipedia to get the movie,
otherwise moves on. Need to find a clearer way of writing the data as well
as a printout at the that summarizes the overall  run like what movies
failed and what movies didnt.